Refactor package structure: move `adapter.web` classes to `adapter.webapi` for consistency across modules
