--- src/main/java/live/b/api/area/adapter/web/AreaResponse.java
@@ -1,8 +1,0 @@
-package live.b.api.area.adapter.web;
-
-public record AreaResponse(
-        String areaName,
-        String imageUrl
-) {
-
-}
