+++ src/main/java/live/b/api/area/adapter/webapi/AreaResponse.java
@@ -1,0 +1,8 @@
+package live.b.api.area.adapter.webapi;
+
+public record AreaResponse(
+        String areaName,
+        String imageUrl
+) {
+
+}
