+++ src/main/java/live/b/api/course/adapter/webapi/PopularCourseResponse.java
@@ -1,0 +1,19 @@
+package live.b.api.course.adapter.webapi;
+
+import java.util.List;
+
+public record PopularCourseResponse(
+        Long courseId,
+        String courseName,
+        String courseTitle,
+        String courseDescription,
+        Float durationHours,
+        String courseImageUrl,
+        List<SpotResponse> spots
+) {
+    public record SpotResponse(
+            Long spotId,
+            String spotName
+    ) {
+    }
+}
