--- src/main/java/live/b/api/event/adapter/web/EventResponse.java
@@ -1,9 +1,0 @@
-package live.b.api.event.adapter.web;
-
-public record EventResponse(
-        Long eventId,
-        String eventCategory,
-        String eventImageUrl,
-        String eventName
-) {
-}
