+++ src/main/java/live/b/api/event/adapter/webapi/EventResponse.java
@@ -1,0 +1,9 @@
+package live.b.api.event.adapter.webapi;
+
+public record EventResponse(
+        Long eventId,
+        String eventCategory,
+        String eventImageUrl,
+        String eventName
+) {
+}
