--- src/main/java/live/b/api/festival/adapter/web/FestivalResponse.java
@@ -1,7 +1,0 @@
-package live.b.api.festival.adapter.web;
-
-public record FestivalResponse(
-        Long festivalId,
-        String festivalImageUrl
-) {
-}
