+++ src/main/java/live/b/api/festival/adapter/webapi/FestivalResponse.java
@@ -1,0 +1,7 @@
+package live.b.api.festival.adapter.webapi;
+
+public record FestivalResponse(
+        Long festivalId,
+        String festivalImageUrl
+) {
+}
