--- src/main/java/live/b/api/notice/adapter/web/NoticeResponse.java
@@ -1,4 +1,0 @@
-package live.b.api.notice.adapter.web;
-
-public record NoticeResponse() {
-}
