+++ src/main/java/live/b/api/notice/adapter/webapi/NoticeResponse.java
@@ -1,0 +1,4 @@
+package live.b.api.notice.adapter.webapi;
+
+public record NoticeResponse() {
+}
