--- src/main/java/live/b/api/spot/adapter/web/PopularSpotResponse.java
@@ -1,8 +1,0 @@
-package live.b.api.spot.adapter.web;
-
-public record PopularSpotResponse(
-        Long spotId,
-        String spotName,
-        String spotImageUrl
-) {
-}
