+++ src/main/java/live/b/api/spot/adapter/webapi/PopularSpotResponse.java
@@ -1,0 +1,8 @@
+package live.b.api.spot.adapter.webapi;
+
+public record PopularSpotResponse(
+        Long spotId,
+        String spotName,
+        String spotImageUrl
+) {
+}
