--- src/main/java/live/b/api/theme/adapter/web/ThemeResponse.java
@@ -1,8 +1,0 @@
-package live.b.api.theme.adapter.web;
-
-public record ThemeResponse(
-        String themeName,
-        String themeImageUrl,
-        String themeImageName
-) {
-}
