+++ src/main/java/live/b/api/theme/adapter/webapi/ThemeResponse.java
@@ -1,0 +1,8 @@
+package live.b.api.theme.adapter.webapi;
+
+public record ThemeResponse(
+        String themeName,
+        String themeImageUrl,
+        String themeImageName
+) {
+}
