--- src/main/java/live/b/api/weather/adapter/web/WeatherApi.java
@@ -1,16 +1,0 @@
-package live.b.api.weather.adapter.web;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/api/v1/weather")
-public class WeatherApi {
-
-    @GetMapping
-    public ResponseEntity<WeatherResponse> weather() {
-        return ResponseEntity.ok(new WeatherResponse("23"));
-    }
-}
