--- src/main/java/live/b/api/weather/adapter/web/WeatherResponse.java
@@ -1,6 +1,0 @@
-package live.b.api.weather.adapter.web;
-
-public record WeatherResponse(
-        String temperature
-) {
-}
