+++ src/main/java/live/b/api/weather/adapter/webapi/WeatherApi.java
@@ -1,0 +1,16 @@
+package live.b.api.weather.adapter.webapi;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/api/v1/weather")
+public class WeatherApi {
+
+    @GetMapping
+    public ResponseEntity<WeatherResponse> weather() {
+        return ResponseEntity.ok(new WeatherResponse("23"));
+    }
+}
