+++ src/main/java/live/b/api/weather/adapter/webapi/WeatherResponse.java
@@ -1,0 +1,6 @@
+package live.b.api.weather.adapter.webapi;
+
+public record WeatherResponse(
+        String temperature
+) {
+}
